dashboard changes (images added to CDN)

+++ app/api/users/route.ts
@@ -1,3 +1,4 @@
+import getCurrentUser from "@/app/actions/getCurrentUser";
 import prisma from "@/app/libs/prismadb";
 import bcryptjs from "bcryptjs";
 import { NextResponse } from "next/server";
@@ -40,12 +41,23 @@ export async function PUT(request: Request) {
 }
 
 export async function DELETE(request: Request) {
-  const { id } = await request.json();
-  const deletedUser = await prisma.user.delete({
-    where: { id },
-  });
+  try {
+    const { id } = await request.json();
+    const currentUser = await getCurrentUser();
+    if (currentUser == null) {
+      return NextResponse.error();
+    }
+    if (currentUser.userRole === "ADMIN" && id === currentUser.id) {
+      return NextResponse.error();
+    }
+    const deletedUser = await prisma.user.delete({
+      where: { id },
+    });
 
-  return NextResponse.json(deletedUser);
+    return NextResponse.json(deletedUser);
+  } catch (error) {
+    return NextResponse.error();
+  }
 }
 
 export async function GET(request: Request) {
